1.5 Bugfix
Vanilla rimworld now checks Pawn_InteractionsTracker.InteractedTooRecentlyToInteract twice now. This is a redundant check but a very minor issue so i haven't mentioned it to the devs.

Simply subtracting the number checked against in the mentioned method was an easier fix than overwriting the entire methods or making a transpiler.